TTS/requirements.txt
celery==5.3.5
redis==5.0.1
gTTS==2.5.0
pyttsx3==2.90
TTS==0.22.0
torchaudio==2.1.0
PyYAML